Move reservation series serializers to different modules in a package
